+++ apps/docs/terraform/main.tf
@@ -1,0 +1,19 @@
+terraform {
+  required_providers {
+    aws = {
+      source  = "hashicorp/aws"
+      version = "~> 5.34"
+    }
+  }
+
+  required_version = ">= 1.7.0"
+}
+
+provider "aws" {
+  region = var.aws_region
+}
+
+provider "aws" {
+  alias  = "us-east-1"
+  region = "us-east-1"
+}