+++ libs/auth/src/lib/services/token.service.ts
@@ -75,7 +75,7 @@ export class TokenService<
         secret,
       });
     } catch (error) {
-      this.logger.error(`Error while trying to verify a token`, error);
+      this.logger.error({ err: error }, `Error while trying to verify a token`);
 
       throw new GeneralUnauthorizedException(undefined, error);
     }