+++ libs/i18n/src/lib/utils/type-guards.ts
@@ -1,0 +1,7 @@
+import { I18nOptionResolver, ResolverWithOptions } from '../interfaces';
+
+export const isResolverWithOptions = (
+  resolver: I18nOptionResolver,
+): resolver is ResolverWithOptions => {
+  return 'use' in resolver;
+};