+++ libs/mail/src/lib/config/sendgrid/sendgrid.config.ts
@@ -1,0 +1,16 @@
+import { IsEmailLocalized, IsStringLocalized } from '@softkit/validation';
+import { IsOptional } from 'class-validator';
+import { GeneralConfig } from '../general.config';
+
+export class SendgridConfig extends GeneralConfig {
+  @IsStringLocalized()
+  apiKey!: string;
+
+  @IsStringLocalized()
+  @IsEmailLocalized()
+  defaultFromEmail!: string;
+
+  @IsStringLocalized()
+  @IsOptional()
+  defaultFromName?: string;
+}