+++ libs/mail/src/lib/services/index.ts
@@ -1,3 +1,4 @@
 export * from './abstract-mail.service';
-export * from './mailgun.service';
-export * from './vo';
+export * from './mailgun/mailgun.service';
+export * from './mailgun/vo';
+export * from './sendgrid/sendgrid.service';