+++ libs/file-storage/CHANGELOG.md
@@ -1,4 +1,26 @@
 Softkit Core Libraries Changelog
+## [0.1.8](https://github.com/softkitit/softkit-core/compare/file-storage-0.1.7...file-storage-0.1.8) (2024-03-06)
+
+
+### Bug Fixes
+
+* **file-storage:** added s3 file service to exports list ([78ae619](https://github.com/softkitit/softkit-core/commit/78ae61965fd08958e310f64b6f9fdf1c01612a38))
+
+## [0.1.7](https://github.com/softkitit/softkit-core/compare/file-storage-0.1.6...file-storage-0.1.7) (2024-03-06)
+
+
+### Bug Fixes
+
+* **file-storage:** updated the package json ([8ae4397](https://github.com/softkitit/softkit-core/commit/8ae43976b9ea16f6341afc4ff4f11e2945dceaea))
+
+## [0.1.6](https://github.com/softkitit/softkit-core/compare/file-storage-0.1.5...file-storage-0.1.6) (2024-03-06)
+
+
+### Bug Fixes
+
+* **file-storage:** updated the default file providers ([c2f6165](https://github.com/softkitit/softkit-core/commit/c2f616572072bb9bf1c0ff0c36f4dff288d057f4))
+* **file-storage:** updated the default file providers ([1d4cd8c](https://github.com/softkitit/softkit-core/commit/1d4cd8c7fc926d7083f10ab616d6a3a5a27116c2))
+
 ## [0.1.5](https://github.com/softkitit/softkit-core/compare/file-storage-0.1.4...file-storage-0.1.5) (2024-02-05)