+++ libs/mail/CHANGELOG.md
@@ -1,4 +1,38 @@
 Softkit Core Libraries Changelog
+## [0.3.2](https://github.com/softkitit/softkit-core/compare/mail-0.3.1...mail-0.3.2) (2024-05-28)
+
+
+### Bug Fixes
+
+* **mail:** added more tests for sendgrid ([b87384b](https://github.com/softkitit/softkit-core/commit/b87384b89569d03e88f16a4fa90f81a73dae994a))
+
+## [0.3.1](https://github.com/softkitit/softkit-core/compare/mail-0.3.0...mail-0.3.1) (2024-05-24)
+
+
+### Bug Fixes
+
+* **mail:** added optional instalation for the sendgrid library and mailgun library ([0b5eb33](https://github.com/softkitit/softkit-core/commit/0b5eb33e858382d99470845cacffe4292abdfe7b))
+
+## [0.3.0](https://github.com/softkitit/softkit-core/compare/mail-0.2.5...mail-0.3.0) (2024-05-23)
+
+
+### Features
+
+* **mail:** added send grid library integration ([a7deef1](https://github.com/softkitit/softkit-core/commit/a7deef1f6fc5b7a520447d3ce3900c0705f10def))
+
+
+### Bug Fixes
+
+* **mail:** small tests adjustments ([913ff77](https://github.com/softkitit/softkit-core/commit/913ff77fea0a542ba7d982deacc37de1f671276a))
+* **mail:** temporary lover coverage threshold ([ca5422b](https://github.com/softkitit/softkit-core/commit/ca5422b04b570d5307fffd79ca4e1fb19ea3f0e3))
+
+## [0.2.5](https://github.com/softkitit/softkit-core/compare/mail-0.2.4...mail-0.2.5) (2024-05-17)
+
+
+### Bug Fixes
+
+* **mail:** improved logs to include context and show errors stacktrace properly ([748a0cf](https://github.com/softkitit/softkit-core/commit/748a0cfc23bd8409760818cf52ea4bb554ab9097))
+
 ## [0.2.4](https://github.com/softkitit/softkit-core/compare/mail-0.2.3...mail-0.2.4) (2024-02-05)
 
 ## [0.2.3](https://github.com/softkitit/softkit-core/compare/mail-0.2.2...mail-0.2.3) (2024-01-26)